+++ packages/broker/test/integration/local-propagation.test.ts
@@ -26,7 +26,6 @@ describe('local propagation', () => {
         brokerWallet = new Wallet(await fetchPrivateKeyWithGas())
 
         broker = await startBroker({
-            name: 'broker1',
             privateKey: brokerWallet.privateKey,
             trackerPort,
             httpPort,