+++ packages/client/src/index-exports.ts
@@ -19,7 +19,6 @@ export {
     UserPermissionAssignment,
     PublicPermissionAssignment
 } from './permission'
-export { UserDetails } from './LoginEndpoints'
 export { StreamValidationInfo } from './StreamEndpoints'
 export { StorageNodeAssignmentEvent, StorageNodeMetadata } from './StorageNodeRegistry'
 export { SearchStreamsPermissionFilter } from './searchStreams'
@@ -28,7 +27,6 @@ export {
     StrictStreamrClientConfig,
     SubscribeConfig,
     ConnectionConfig,
-    DataUnionConfig,
     TrackerRegistrySmartContract,
     NetworkConfig,
     DebugConfig,
@@ -47,7 +45,6 @@ export {
     ProviderAuthConfig,
     ProviderConfig,
     PrivateKeyAuthConfig,
-    SessionTokenAuthConfig,
     XOR,
     Without
 } from './Ethereum'
@@ -57,7 +54,6 @@ export { UpdateEncryptionKeyOptions } from './encryption/GroupKeyStoreFactory'
 
 export { ConfigTest } from './ConfigTest'
 export { NetworkNodeStub } from './BrubeckNode'
-export * from './dataunion/DataUnion'
 export { NotFoundError, ErrorCode } from './authFetch'
 export { SignalListener } from './utils/Signal'
 export * from './types'