+++ packages/network/src/createNetworkNode.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid'
-import { MetricsContext } from './helpers/MetricsContext'
+import { MetricsContext } from './helpers/Metric'
 
 import { TrackerInfo, AbstractNodeOptions } from './identifiers'
 import { NodeToTracker } from './protocol/NodeToTracker'
@@ -28,7 +28,6 @@ export interface NetworkNodeOptions extends AbstractNodeOptions {
 
 export const createNetworkNode = ({
     id = uuidv4(),
-    name,
     location,
     trackers,
     metricsContext = new MetricsContext(),
@@ -44,7 +43,7 @@ export const createNetworkNode = ({
     webrtcDisallowPrivateAddresses = true,
     acceptProxyConnections
 }: NetworkNodeOptions): NetworkNode => {
-    const peerInfo = PeerInfo.newNode(id, name, undefined, undefined, location)
+    const peerInfo = PeerInfo.newNode(id, undefined, undefined, location)
     const endpoint = new NodeClientWsEndpoint(peerInfo, trackerPingInterval)
     const nodeToTracker = new NodeToTracker(endpoint)